requirements.txt
# Core ML dependencies
ultralytics>=8.0.0
opencv-python>=4.5.0
easyocr>=1.6.0

# Data handling
pandas>=1.3.0
numpy>=1.21.0
openpyxl>=3.0.0

# Dataset management
roboflow>=1.0.0

# Web interface
flask>=2.0.0
werkzeug>=2.0.0

# Environment
python-dotenv>=0.19.0

# Optional: GPU support
# torch>=2.0.0
# torchvision>=0.15.0
